pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "email-validator>=2.2.0",
    "flask-migrate>=4.1.0",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "google-api-core>=2.24.1",
    "google-generativeai>=0.8.4",
    "openai>=1.61.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "python-telegram-bot[rate-limiter]==20.7",
    "ratelimit>=2.2.1",
    "requests>=2.32.3",
    "slack-sdk>=3.34.0",
    "sqlalchemy>=2.0.38",
    "telegram>=0.0.1",
    "tenacity>=9.0.0",
    "twilio>=9.4.5",
    "werkzeug>=3.1.3",
    "python-dotenv>=1.0.1",
    "alembic>=1.14.1",
    "flask-wtf>=1.2.2",
]
